use generics for rules api
